pyproject.toml
[project]
name = "gpt-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "python-telegram-bot>=21.11.1",
    "requests>=2.32.3",
]